version3/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.1

# Diffusion model dependencies
diffusers>=0.19.0
invisible-watermark>=0.1.5
xformers>=0.0.20
bitsandbytes>=0.41.0
peft>=0.5.0

# Image processing and segmentation
opencv-python>=4.7.0
segment-anything>=1.0
nibabel>=5.0.0  # For NIfTI files (MRI)
scikit-image>=0.20.0
scikit-learn>=1.2.0

# MONAI (for medical image specific processing, optional)
monai>=1.2.0

# Utilities
tensorboard>=2.12.0
wandb>=0.15.0  # For experiment tracking (optional)
jsonschema>=4.17.0
pycocotools>=2.0.6
omegaconf>=2.3.0
PyYAML>=6.0
pytest>=7.3.1  # For testing

# SAM2 specific dependencies 
timm>=0.9.2  # Vision Transformer components
simple-parsing>=0.1.4  # For SAM2 config parsing
fairscale>=0.4.13  # For SAM2 model scaling
termcolor>=2.3.0  # For colored terminal output in SAM2
setuptools>=68.0.0  # Required by SAM2 setup

# Additional image utils
gradio>=3.32.0  # For demos (optional)
einops>=0.6.0  # Array dimension operations
albumentations>=1.3.0  # Image augmentations 